DE759 - changed default section number from 0 to whatever the default position of that section is in the customer journey
After discussion with acting BA validator he is happy with keeping the section title
(Section 6 of 10 for example) on the /employment/employment page, but having it default
to section 6 if the user selects 'no' to both employment and self-employment.
This will stop it showing 'Section 0 of 8'. However, the underlying defect will remain,
as it is caused by this page belonging to one of the employment sections,
therefore saying no to both means the page they are on is no longer part of the journey
they are on. The only real fix is to remove section numbering from this page or to make
this page its own section. Neither of these solutions were acceptable to acting BA or
the product manager.