Fixing the deploy to lab for c3.
This fixes the deploy to lab by turning the Scala deploy to lab into a
similar process to the Java apps.

In the background the lab salt stack has been updated to put a deploy.sh
into place for the scala users and the jenkins deploy key has been added
to each user.